requirements.txt
numpy==1.26.0
torch
torch-scatter
torch-sparse
pytorch_geometric
